Make Applet config window a tiny bit smaller and reduce minimum size
There's no point in setting an implicit size that's the same as the
minimum size so let's reduce the minimum size to be smaller than the
minimum size so that the window can be made smaller by the user, or by
the system in case a small screen is being used. Also reduce the default
height a tiny bit.

BUG: 443712
FIXED-IN: 5.24
(cherry picked from commit 4e13d08c54d79bb8983f47ff77065d4e683c8c05)