Fix create-folder-shortcut (F10)
The create folder shortcut worked as long as the user did not right-
click into an empty space on the desktop to invoke the context menu.

The problem was that the create-folder-action was set to not visible
so that it is not shown in the context menu. But as soon as the context
menu is opened, the action is set to not visible in the plasmoid
contextual action list. The reason it works before opening the menu is
because the action for the plasmoid is created without setting it to
not visible.

This commit fixes the bug by using an already existing workaround class
that uses an event filter to respond to the
KStandardShortcut::createFolder key sequence.

BUG: 443512


(cherry picked from commit 12350fb1242c4dbfdb547926bdb093eb8da96b4b)