emojier: Show placeholder message instead of disabling recents when empty
Right now we disable the recents view when it's empty, but this relies
on querying recentEmojiModel.count, which is not a property, so it
doesn't update as the model fills up with recents. Making it a property
is annoying, and a better UX is simply to show a placeholder message on
the page when it's empty as we do in other places. The code already
skips showing the view on launch if it's empty, so everything still
works.

BUG: 443972
FIXED-IN: 5.24