Edit mode toolbar: Replace Activity Switcher button with KScreen KCM button
The Activity Switcher has nothing to do with editing global settings, so
it is not an appropriate thing to be on this toolbar.

Let's replace it with a button to open the KScreen KCM, which does
involve editing global things, and letting people access it from here
may help in cases where something's gone wrong and the screen with their
panel isn't showing it anymore so they can't access a means to launch
apps to get to System Setitngs to get to the KScreen KCM to juggle
things around and try to fix the problem.

CCBUG: 355679