Remove middle-click-paste actions from panel config files
In the past, panels were configured to add a note on middle-click. This was changed
several years ago in
https://invent.kde.org/plasma/plasma-desktop/-/commit/f596c2cb6fa79ad311149ac20d142380d739c15f,
but the "MiddleButton;NoModifier=org.kde.paste" action was never removed from users'
config files, so some people still got this undesirable behavior with no GUI method
to change it.

This commit adds an update script that removes the config entry that enables this
behavior for all of the user's panels.

BUG: 425852
FIXED-IN: 5.24