[Folder View] Add another large icon size
The current set of icon sizes has three fairly small sizes, but only
two reasonably large ones. Arguably only the largest size (128px) is
truly "large" given the size of typical displays today. This makes the
size distribution a bit unbalanced, and impairs the use case of wanting
to have only a small number of very large icons on the desktop for
display, presentation, or user assistance purposes.

Accordingly, this commit adds another large icon size to the top end of
the options: enormous*2, which is 256px in practice.

No config migration is needed since existing icon size steps are not
changed.