Replace Kate with KWrite in favorites
Kate is a programmer's text editor that may not be the best choice for a
general-purpose note taking or text editing tool. Furthermore, its name
and icon do not make its purpose obvious, and this is important for the
Favorites view, where any given menu representation is not guaranteed to
show the description (and in fact most do not) so being recognizable by
icon or name is important.