taskmanager: Use default WindowType in GroupDialog
Before frameworks/plasma-framework!391, PopupMenu only has the default
role, which blocks a dialog with the PopupMenu role from receiving
any QFocusEvent on Wayland.

In case frameworks/plasma-framework!391 cannot be merged, this commit
sets GroupDialog as a Normal winodw, so it's possible for GroupDialog to
receive QFocusEvent normally, and the behavior on X11 can remain the same.

BUG: 444328