kcms/runners: Rename to "Plasma Search" and clarify scope of options
In the past this KCM was called "Plasma Search", but we renamed it to
"KRunner" because its options only applied to KRunner (and technically
the Milou plasmoid but we forgot about that).

Then we made Kickoff use the same plugins as KRunner. Then we added the
Overview effect which also has a KRunner-powered search that uses the
same plugins.

So it's time to return the KCM to its former name "Plasma Search" since
this is now more accurate. One slight complication is that since then
the KCM gained a few options that are in fact specific only to KRunner,
so they are slightly renamed to indicate this.

In the future it might be better to move these into a KRunner-specific
config window like we have for Plasma applets.