containments/panel: Fix state when dragging gets canceled
In panel edition, when dragging an applet to reorder the layout,
we were not handling the case when the dragging operation gets
canceled.

Dragging cancelation can be triggered by pressing <Esc> or other
focus-intrusive operations like a delayed screenshot.

This led to an unconsistent state of the panel where the dragged
applet would just stay out the layout, floating in the last
position it was dragged to.

For now, acting on the `canceled` signal just as if the dragging
was `released` seems the simpler to avoid a messed-up panel.

BUG: 446814