Break Toolbox into two rows
The current ToolBox has too many items and ends up often overflowing the screen,
rendering labels unreadable, and in some cases, items inaccessible.

This is a quick-fix commit to ensure that the toolbox doesn't exceed
screen bounds for more cases, as a 'proper' fix of wrapping items based
on available space is exceedingly difficult to implement with the
architecture of the toolbox.