[AppletError] Improve visuals and make it more convenient to use
You can now see the error directly from AppletError if there's enough
room. If there isn't enough room, you will be given a button to open the
error in a new window.

The error window is now able to appear above pinned plasmoids so that
you
don't need to move the window to read it.

The error now uses the system's monospace font.

AppletError no longer has warnings when it is loaded. It's kind of
ironic for an error dialog to have errors or warnings.