[kcms/keyboard] Fix populating layouts when reloading config
Previously, it populated the internal list of layouts before reloading
settings from kcfg. That lead to KCMKeyboardWidget::configureLayoutsChanged
getting called for ticking the checkbox while the list of layouts was
already nonempty, which is an invalid/unexpected state. In its confusion it
cleared the list.

Just load kcfg settings first to avoid this.

BUG: 448381