Update click method items in load()
This makes sure we re-read the values from the backend. While these
properties are marked as constant, they apparently change after
LibinputTouchpad::getConfig is called, causing issues because the
changed values aren't re-read.

While I would love to fix the root cause (the properties changing while
being marked constant) it would be a massively larger undertaking as in
my opinion the entire code of this should be completely overhauled,
right now it is fraught with races due to all the manual state
management and lack of proper bindable properties.

BUG: 447105