foldermodel: hide paste action when right-clicking on files
It doesn't make sense to offer a paste action when right-clicking on a
file as it implies that the paste will overwrite the file, or at least
interact with it in some way, but this is false.

Dolphin follows this logic and doesn't offer a paste action when
right-clicking on files, but Folder View does not, and it should.
This commit makes the Paste action not appear if the thing that was
right-clicked on is a file, rather than a folder or the wallpaper.

BUG: 448913
FIXED-IN: 5.24