kcms/touchpad: Remove weird Q_EMIT changed(false) in resizeEvent
That looks rather out of place there and does actually disable the apply button
when resizing the window. I don't understand what it's supposed to be doing
there, but if it has an actual purpose it needs to be addressed differently.

CCBUG: 449843


(cherry picked from commit 369eab6457965de99119f9966a009f98a60745bc)