kcm/keys: Support adding arbitrary applications
KOpenWithDialog actually has a feature where it will create a desktop
file if we specify an application that doesn't have one. With that
enabled, you can add any application to the list of applications to
launch.

Unfortunately, in that case, we get an absolute path to the desktop file
instead of just the filename. So we need to check if that is the case
and if so, only use the filename, since the rest of the code assumes
we're dealing with filenames instead of full paths.