applets/margins-separator: make smaller normally, and more visible in Edit mode
Margin separator by default has a large width which may not be desirable always
and it is also not directly related to its core functionality. Now it takes up
only 1px of space (the minimum amount) outside of Edit Mode, but expands and
becomes more visually obvious in Edit Mode.

BUG: 451164
BUG: 433475
FIXED-IN: 5.25