Folder View: Make popup dialog wide enough for one more grid cell
Right now it is 2 pixels too narrow, so there is always a huge gap on
the right side. Use a standard Unit that knows how to adjust with the
font size to avoid this, in lieu of being able to get the scrollbar
width.

BUG: 417539
FIXED-IN: 5.24.5
(cherry picked from commit 2ebde10325ac9077469210a4cc7fa89480113053)