applets/taskmanager: increase spacing in tablet mode
Now that we have the ability to manually configure IOTM task spacing,
let's make use of this feature to automatically set it to its highest
spacing value while in Tablet Mode to improve touchability, just like we
already do for System Tray icons.

As a part of this, the Huge size is removed and the Large size is made
to do the same thing as the Huge size, so that we have fewer options and
those options are 100% consistent with what we show in the System Tray.