Fix the busy indicator not being centered properly
The busy indicator in the task manager does not take into account the amount
of margins and may look bad in larger spacing between icons. This issue seems
to be there out always but it is more apparent while using a larger spacing
between icons. The indicator should shift itself half the amount of margin
left or bottom depending on the panel type.