automount: Disable back automount by default
This setting was enabled to make the KCM defaults slightly more consistent, as
the partial settings `AutomountOnLogin` and `AutomountOnAttached` are enabled by
default and the KCM enables the general `AutomountEnabled` settings from those.

But this seems to be problematic for some use cases (reported on fresh-installed
systems like Fedora), so we need to disable it by default.

As the KCM compose the general option from the two partial ones, we need to disable
them too, so they're in sync and the Default state is stable.


(cherry picked from commit f251558c6946497049f6bd101ccf7dff4d082178)