Implement optional floating panel for Plasma Themes
This implements a way to make plasma panels floating. It is opt-in.

Fitts' law is preserved; clicking on any empty space between the panel and the
screen near it will redirect input inside the panel. When any windows are
maximized, the panel de-floats by filling the available space with the svg and
moving the content down by `relevant margin / 2` so that it doesn't move from 
its position when the panel is floating.

Known visual quirks:
- Shadows are not supported for floating panels
- Open applets extend over panel ends
- When a floating panel de-floats because of maximized windows, its widgets will
  not touch the edges of the panel only alternative to this would be to break
  applets' spatial positioning when their panel de-floats, which would be worse.