Change default keyboard switching shortcut to Meta+Alt+K
The current shortcut is Ctrl+Alt+K which would conceivably override app
shortcuts, since it doesn't use the Meta key.

This is safe to change in the code rather than adding a new one and
keeping the old one around for compatibility's sake because
KGlobalAccel::setShortcut by default preserves old shortcuts when they
are changed, and we aren't overriding that behavior here. So only new
installs will get the new shortcut and existing user installs will be
untouched.