Make windowlist app name display configurable
Before Plasma 5.25, this widget's panel button was always just an icon. That
changed in 5.25, so now it also has text when used on a horizontal panel,
where there's space for it. This makes the widget awkward if there are other
widgets adjacent to it, because its width will constantly change based on the
name of the active application, causing other widgets to shift around and
breaking the user's muscle memory. On the other hand, this is fine as long as
the widget has a flexible spacer between it and any other widgets, or if its
adjacent widgets have their own built-in space-filling flexible spacers, as
the Task Manager and Global Menu widgets do.

Because that can't be programmatically determined, and also because users of
horizontal panels may prefer the icons-only version anyway, let's make this
configurable.

This new config option only works when the widget is located on a horizontal
panel, so it automatically disables itself and shows an explanatory message
when it's being used in any other circumstance.

| Config UI when the widget is on a horizontal panel | Config UI when the widget is on a vertical panel | Config UI when the widget is on the Desktop |
| -------------------------------------------------- | ------------------------------------------------ | ------------------------ |
| ![Horizontal_panel](/uploads/a8b6da1269f316dd8eec024ada89b1e6/Horizontal_panel.png) | ![Vertical_panel](/uploads/f079d4e34f67df93e4cdca923c68c29f/Vertical_panel.png) | ![Desktop](/uploads/3aed5d892fe4f8d83cde6a8fddef476c/Desktop.png) |

BUG: 455560
FIXED-IN: 5.26

@cblack @teams/usability @teams/vdg