Remove FolderModel::openSelected()
This function gets run when you select multiple items and right-click >
Open. However there is already a FolderModel::runSelected() function
that does basically the same thing, but with more sophisticated handling
of the selected files. So let's use that instead and delete
openSelected() to remove a redundant parallel code path.