ShowDesktop: Use the new DBus API for showing the desktop
This unifies the code paths (underneath KWindowSystem goes through two
separate code paths) and generally makes the code a lot more direct
instead of going through a bunch of layers. Additionally, it allows the
KWin-side of things to check if the caller disappears and quit the "show
desktop" mode if that happens.

BUG: 449445