applets/kickoff: add "compact list item style" option and use new tall style by default
Kickoff uses an unusual way of locating list items' descriptions: it
puts them on the same line as the title, aligned to the right.

This saves vertical space but suffers from several issues:

- Because both labels need to inhabit the same line, a long title will
  cause the description to be elided into uselessness
- It's a different style than what we use for other list items with
  fundamentally the same set of data; we don't re-use users' existing
  familiarity with KDE-style list items that typically have the caption,
  subtitle, or description below the label/title
- Saving vertical space is not a win here because we want the applet to
  be touch-friendly, and short list items make that worse, not better

Accordingly, this MR adds an option to switch the list items between two
styles:
- The current style, which is now called "compact list item style"
- A new default style with the subtitle below the main text, which is more
  conventional and more touch friendly.

This new default list item style is intended to improve the "Gwenview? What's
a Gwenview?" effect that users have been complaining about. Despite the fact
that the compact/inline list items have a caption that actually tells you what
branded apps like Gwenview are, people seem blind to it because they don't
expect to see a description in that location. No other KDE list items do this
and it's a fairly unusual style outside of KDE as well.

When using Touch mode, compact mode is disabled if it was manually activated.


| New default setting (tall list items) | compact list items | Config window now |
| ------------------------------------- | ------------------ | ----------------- |
| ![Tall_style](/uploads/9af537fbbabe461aff7a170139b8095f/Tall_style.png) | ![compact_mode](/uploads/ce119b5fb345f2b784bbbba9e73a2889/compact_mode.png) | ![Config_window](/uploads/2204db79fbdf06f653d1ae968a472424/Config_window.png) |

BUG: 454455
FIXED-IN: 5.26