Disable Drag and Drop for touch, fix drag and drop icons
Drag and drop conflicted with drag/flick scrolling and drag and drop was
deemed less important than scrolling or opening the context menu for
touch. Maybe in the future we can have a way to at least rearrange
favorites with touch input.

Drag and drop icons seem to have been broken by commit
0868a5ae8154d10ccc7488034bc5b6e458a17968 and are now fixed.

BUG: 450448