applets/kickoff: fix "ReferenceError: plasmoid is not defined" error
This error happened because KickoffListView was asking for metrics that
come from a KickoffListItem living in KickoffSingleton. However that
item needs access to a config property to determine its own size, and
singletons can't access per-applet config data, because they're shared
across all instances of the app.

This commit fixes the issue by moving the metrics object for the
variable-size list item from the singleton to the root item of the
applet instance.