Disable Drag and Drop for touch, fix drag and drop icons
Drag and drop conflicted with drag/flick scrolling and drag and drop was
deemed less important than scrolling or opening the context menu for
touch. Maybe in the future we can have a way to at least rearrange
favorites with touch input.

BUG: 450448

This commit does the same thing as
66507b5df163071128a90983d10f699aa1a9b3cf, but that one couldn't be
cherry-picked.