Fix KCM duplicates in landingpage
In systemsettings, we no longer write a ".desktop" at the end of the KCM ids.
This change was made during Plasma 5.24. However, it causes duplicates on some setups.
To avoid this, we ignore all of the old .desktop URLs. To ensure that we fill all of the 6 available
items, the limit is increased to 12 and we check against a list how many results we have ignored.

BUG: 449563


(cherry picked from commit 9a78c778c24564bf71571c3993deddb20baa104f)