panelconfiguration: use radio buttons for mututally exclusive items
Previously we used conjoined buttons, but in the PC3 port to be
compatible with Qt6, we lost the conjoined appearance, making the
options look a lot less mutually exclusive than they used to.

To remedy this, instead use Radio Buttons, which are the most original
and recognizable UI controls to denote mutually exclusive actions.