applets/keyboardlayout: re-introduce "label+flag" display style option
This feature got lost in the QML port and is re-added here.

The old version of the feature drew the label on top of the flag with no
background, thereby making the label appear invisible for certain flags.
To alleviate this, the new implementation draws the label on a badge in
the corner, like the Battery & Brightness widget does.

Because this changes the format of the setting that controls the display
style (going from a bool to an enum), an update script is provided to
migrate people's old settings as needed.

BUG: 444864
FIXED-IN: 5.26