kcms/ksmserver: Remove UI to choose a "default logout screen option"
The concept of the "default logout screen option" doesn't exist anymore
in most places. The only thing that still respects the value of this
setting is when you press the physical power button to show the logout
screen, when using the Powerdevil option to show the logout screen when
pressing the power button. This is a rather esoteric and niche use
case, and showing a UI here to configure what default button is
highlighted for that use case is quite heavy, misleading users into
thinking that this seetting affects more than it really does. Let's
remove it to simplify this KCM.