fix: Keyboard layout shortcuts activate wrong language
Fixes spare layout's shortcuts activating wrong layout after first use, see the bug for the details.

When we switch to a spare layout (by the shortcuts or via applet's menu), rearrangement of layouts happen - the spare one become the last in the main loop, preempting the previous one on it's place.  
Layouts-specific shortcut actions activate the layout by index which is hold in the action's data. Thus, the indexes should be adjusted accordingly.
 
BUG: 444569