kcms/landingpage: Use global themes' real names in quick chooser
By using generic "Light" and "Dark" terminology here, we're giving
people the mistaken impression that the global theme buttons on the
Quick Settings page comprise a generic light/dark toggle switch. But
they do not; they are simply quick links to the Breeze Light and Breeze
Dark global themes. So let's clarify that by using the actual names of
whatever global themes are configured to appear here.

BUG: 459367
FIXED-IN: 5.26