AppletConfiguration: add "Report a Bug" button to config window
When an applet's JSON metadata defines the bugReportUrl key, its
Configuration page will now display a "Report a Bug" button that the
user can click on to quickly and easily file a bug. This should make it
easier for people to file targeted bug reports that are automatically
filed to the right place so they don't need extra work by big triagers
to move them.

If the applet does not define the bugReportUrl key in its JSON metadata
file, it doesn't get this button, and there are no visual changes.

BUG: 370195
FIXED-IN: 5.27