Bump KDE compiler settings level to 5.85
This is needed to enable QStringBuilder, which is needed for QStringView
concatenation code in emojier to compile with Qt 6. This however also
enables stricter string casing options, which is the bulk of these changes.

To minimize the impact this is already using new string literal operators
from Qt 6 in the files most affected by this.