Port emojier away from deprecated I18N_NOOP
This is rather dirty, the category table should hold KLazyLocalizedString
instances rather than QStrings of the untranslated message, and actual
translation should be done by calling toString() on those, to avoid
duplicating the context.

That however isn't easily possible here as the translation happens on the
QML side, after using the untranslated string in a mapping table for icons.
So this has to do for now.