Also delete entry with Notify flag
Otherwise KWin will not know of the changes.
This does not fix
CCBUG:460345
because of the second issue that prevents the event capturing but once that is solved, removing and adding a binding for the same button would not work.

(cherry picked from commit 979de04073155f114cfd40d4001b508a0a998748)