[applets/digitalclock] Fix font size change when migrating from 5.25
When migrating from 5.25, if custom font is set, the migration script
sets fontSize=10, which causes the look of digitalclock to change from
5.25 (where the font autosizes). And the default value of 10 can be too
small for some users.

So now this settings migration script sets fontSize=72 if custom font is
set in 5.25. For panels thinner than 72pt, this should retain 5.25's
autosize behavior.
cons:
1) Too late for users already upgraded to 5.26
2) The config dialog looks a bit strange with a very large sample text.

BUG: 460415