kcms/runners: Fix org.kde.kconfig.notify DBus signal getting wrong plugin id
Because we do not have metadata in most KCMs, we use the metadata from the plugin.
If we do not explicitly define a pluginId, it gets derived from the filename.
Meaning the pluginId of the KCM is different from the pluginId of the runner.

However, when the RunnerManager receives this signal, it looks matches the runner pluginId
and the pluginId of the KCM.

BUG: 461422


(cherry picked from commit b9602cad04e8b3f35585a2bad0223bf89f58dd29)