applets/taskmanager: always allow using tooltips for grouped tasks
Right now you can only use tooltips to switch between grouped tasks if
they are enabled for hover too. This seems like an arbitrary limitation,
and also results in an awkward UI in which you can disable tooltips for
hover and enable them for switching between grouped tasks, but when you
do so, you see a message that says "sorry, this is an invalid
combination of settings!"

Well, let's make it not be an invalid combination of settings! This
commit does just that, allowing a new combination of settings that users
have requested while removing an awkward UI and reducing code
complexity.

BUG: 450798
FIXED-IN: 5.27