applets/kickoff: fix huge icons in category list
Kickoff's category list can have apps in it, and in fact we ship this
way by default, with KHelpCenter below the categories.

A code change in 0868a5ae8154d10ccc7488034bc5b6e458a17968 to add the new
compact list item style altered how icon size is calculated to take into
account the value of `isCategory`. But this is false for category list
items that happen to be apps. As a result, the apps in the category list
inappropriately use the app style and their icons are too big.

To fix this, we should force the smaller icon size and list item style
for apps in the category list, because there's no room to show the
subtitles here.

To make the code clearer, `isCategory` is renamed to
`isCategoryListItem` to indicate its status as communicating that the
item lives in the category list, not whether or not it actually is a
category.

BUG: 460349
FIXED-IN: 5.26.4