kcms/ksmserver: remove "Offer shutdown options" setting
This setting's text implies that it disables the logout screen or
removes all the other options from it, but it does not. In fact it
appears to do nothing. No session-related UI that I or anyone else can
find respects the value set here, making it a setting that seems from
the user's perspective to be completely useless.

Let's remove it.

BUG: 431573
FIXED-IN: 5.27