kcms/landingpage: remove KUserFeedback UI and code
The user is going to have a chance to turn on User Feedback in the new
welcome wizard, which will be released with Plsma 5.27 and autostart for
all users once. Thus, we no longer need to put this setting on the Quick
Settings page to make sure the user sees it. This removes a *lot* of
code.