kcms/keys: Add an "Add Command" button
After reading a good amount of feedback and questions from users, it seems that
it is not clear that the textbox in the "Add Application…" dialog can be used to
add custom commands.

Some examples of user confusion:
- https://old.reddit.com/r/kde/comments/xw8k06/the_new_shortcut_page_in_wayland_is_a_big/
- https://old.reddit.com/r/kde/comments/x0tmr5/custom_keyboard_shortcut_to_run_command/
- https://old.reddit.com/r/kde/comments/xf59gs/how_do_i_create_a_custom_shortcut/
- https://old.reddit.com/r/kde/comments/xf4a3g/global_mappings/
- https://old.reddit.com/r/kde/comments/xe59pe/custom_shortcuts_disappeared_in_system_settings/

This commit adds a separate button to add a command, which opens a dedicated
overlay sheet that can be used to add a custom command, which then lives in a
new "Commands" section in the sidebar.

BUG: 460036
BUG: 440431
FIXED-IN: 5.27