applets/taskmanager: Make 1 stripe the default
By default the taskmanager allows for applications to form up to 2
stripes, meaning 2 rows of icons. It's very easy to end up with 2
stripes when customizing your panel.

However, this is not a thing most users are used to, prefering
applications to only take up to 1 row.

BUG: 460911

add an object name to the base focus scope

this is needed for appium based teststo actually expand the applet