applets/window-list: Add fallback icons
Currently, the Window List widget will show an empty icon when focused on the desktop,
meaning the button becomes invisible when it's set to only show icons (or on a
vertical panel). It also doesn't show any icon if an application refers to an invalid
icon (eg. An icon that can't be found on the disk).

This commit fixes that by adding fallback icons for Plasma and invalid icons.