applets/taskmanager: Fix icon size when spacing is set to large
By applying the 3x margins on "large" spacing to both horizontal and vertical margins,
we're also compressing the icon to be too small by triggering a condition in the
minimum size check. With this commit, switching to "large" while the panel is at 38px
shouldn't change the size of the icon anymore.

BUG: 462097
FIXED-IN: 5.27