applets/taskmanager: don't honor groupPopups setting in IOTM form factor
This setting is only visible for the traditional non-icons-only Task
Manager. However, if you uncheck the setting and then use the
"Alternatives" functionality to switch to the icons-only (IOTM) version,
then it will still be honored and cause inappropriate behavior in the
IOTM.

To fix this, we need to disregard the value of that setting in the code
when using the IOTM form factor. This is what we do for other settings
that are only applicable to one form factor or the other, but it was
missed here.

BUG: 464627
FIXED-IN: 5.27.1


(cherry picked from commit 687a3b6bce9612436e27f9c0a4f55c0b155981b3)