applets/pager: Improve usability of touchpad scrolling
After using the touchpad to scroll at least once scrolling, with the
mouse wheel no longer works reliably.

I tracked the issue down to the fact that `wheelDelta` is no longer at
0 after scrolling with the touchpad, so for example if it was at -20 and
I scroll once with the mouse 120 gets added to it, then the value is 100,
which is neither >= 120 nor <= -120, so the scroll event gets ignored.
So, this patch resets the value of `wheelDelta` to 0 after enough scroll
wheel delta has accumulated to trigger switching desktops at least once.

With this commit, scrolling with the touchpad still works, and scrolling
with the mouse now works as expected.