Folder View: with menu representation, always open with a single click
The single-click/double click setting applies to file and folder views.
Folder View is definitely a file and folder view. However, when Folder
view is in its list/menu representation, it's not *presenting* as a file
or folder view, it's presenting as a menu. And menus always trigger
their items with a single click.

We already do a similar thing by always activating with a single
touchscreen tap, bypassing the user's single/double-click preference. So
there's precedent for having a hardcoded behavior to support certain UX
paradigms.

Hence, in list/menu mode, don't respect the user's setting for
single/double click and always activate items with a single click.

BUG: 420261
FIXED-IN: 6.0