desktoppackage: Fix ScrollBar overlapping configuration category delegates
Second rule of the fighting club: never zero out ScrollView paddings.
They exist for a reason, and that reason is to keep enough space on the
sides for scrollbars, in an RTL-aware manner. Content width being bound
to ScrollView's own availableWidth ensures that horizontal ScrollBar
won't appear just because of content's implicit width; and content's
width in turn bound to ScrollView's contentWidth ensures that it won't
get clipped or overlapped.