desktoppackage: Fix relaying configuration change signals
Fixes no-so-verbose-as-I'd-like-them-to-be errors from QML/JS Engine:

    file:///usr/share/plasma/plasmoids/org.kde.plasma.digitalclock/contents/ui/configTimeZones.qml:254:
    Error: Unable to determine callable overload.  Candidates are:

Yes, it does not print any candidates, let alone mentioning the object
itself and the name of a signal which wasn't found.

The correct way it to pass a closure (function) to connect() calls,
because only that way JavaScript can preserve binding between an object
and its method. See Function.prototype.bind() for an alternative
approach and in-depth description of this problem.