kcms/ksplash: Fix loading screenshots by URL
Remember that in QML land QUrl becomes a `url` object with not much more
than a toString() method. Empty default-constructed QUrl() is equal to
an empty JavaScript string.

Also, let's set all the roles on the special "None" model entry, so that
we won't need the !!magic `|| ""` workarounds in UI code.