applets/taskmanager: avoid updating dragSource when active is false
grabToImage is async, and when an item is grabbed to an image, the drag
operation can be already cancelled. When dragSource is not null,
MouseHandler will always consider the current drag opertion is to move
tasks, which causes the bug.

STR:

1. Drag a file onto a task and hold LMB
2. Press RMB
3. Release RMB
4. Release LMB
5. Drag a file onto a task again

BUG: 466675
FIXED-IN: 5.27.6


(cherry picked from commit 24e0d2dcec56051cdd0b4e2ef22b4bb5e1d2d6d7)