Revert "kcms/baloo: prompt and warn users to reboot after changing settings"
This reverts commit 4afb2b44bb816319c1d64dff1b6c2e7dbaeb3f63.

Baloo already does pick up config changes at runtime, contrary to the
statement of the original commit.

The "bug" noticed by the reporter is just an artifact of the status/progress
reporting.

Claiming "Baloo needs a restart" when it does not puts Baloo in a bad light,
and should be avoided.

CCBUG:462009
BUG:467531