kcms/mouse: Explicit build options for Wayland and X11 backends
KDE prefers explicit CMake options over a conditional build that
depends on which headers are installed on the user's system.
Two new options are introduced to turn the build for the X11
and KWin+Wayland backends on or off separately.

The mouse KCM will now build both backends by default, as before,
but won't silently exclude kcm_mouse from the build if any of its
required xorg-evdev/xorg-libinput/X11_Xi/X11_Xinput dependencies
are missing. Instead, it will loudly fail CMake configuration with
an error and a hint about the existence of the CMake option.