MoreSettingsMenu: remove unnecessary width and height overrides
The layout's implicit width and height are fine here; no need to
override them. Doing so was making the window so tall that it overflowed
off the top of the desktop when using a 1366x768 screen resolution.
Simply removing the height overrides is enough to fix that for the
default bottom panel height, even when also using a second top panel.

BUG: 469421