Panel: fix applet not returning focus after pressing applet shortcut
When applet shortcut is pressed for the first time, panel also gets
activated, but the status will change to RequiresAttentionStatus after
applet also gets activated later. When applet is activated, panel status
is not AcceptingInputStatus but RequiresAttentionStatus, so
unconditionally changing to AcceptingInputStatus when the current status
is not AcceptingInputStatus is wrong.

BUG: 472909
FIXED-IN: 5.27.8