kded/touchpad: Disable touchpad notification sounds
With https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/2106 newly connected devices make sounds. If the user plugs in a mouse and the touchpad gets disabled or vice-versa we'd get 2 sounds, which is not ideal. As Nate commented in https://invent.kde.org/raploz/blue-ocean-sound-theme/-/issues/6#note_729989 manually enabling or disabling the touchpad won't make a sound, however, since that action is explicitly triggered by the user there's no need for additional feedback.

I also noticed the "touchpad enabled" action didn't show a popup the same way the "touchpad disabled" action did, so I changed Action to Popup there as well to make the behavior consistent between both actions. 

I added the spec-compliant sounds but disabled them by default instead. Also, the plug-in/out sounds were inverted, probably to make a disconnect sound when the mouse was disconnected. That's not necessary anymore so we can use the device-added sound for enabling the touchpad and the device-removed for disabling it.